Add files via upload
This shell script, named backup.sh, is designed to create a backup of files from a specified target directory and store the backup in a destination directory. Let's dive into how this script works and its key functionalities:

Usage:
The script expects two arguments in the command line: target_directory_name and destination_directory_name. If the correct number of arguments is not provided, the script displays an error message and exits.

Validation of Directories:
The script checks if both the target_directory_name and destination_directory_name provided as arguments are valid directories. If any of them is not a valid directory path, it displays an error message and exits.

Timestamp and Backup File Name:
The script obtains the current timestamp (currentTS) in the format 'YYYYMMDDHHmmss' using the date command. This timestamp is later used to generate a unique backup file name in the format backup_YYYYMMDDHHmmss.tar.gz.

Displaying Target and Destination Directories:
The script displays the paths of the target and destination directories for clarity.

Finding Files for Backup:
The script searches for files in the target_directory_name that have been modified after yesterday's timestamp. It uses the find command to identify such files.

Filtering and Preparing Files for Backup:
Files found by the find command are filtered to exclude symbolic links and directories. Only regular files (-f) are considered for backup. The file paths are collected in an array toBackup.

Creating the Backup Archive:
The script creates a tar archive named after the current timestamp (backupFileName). It compresses the files listed in the toBackup array into this archive using the tar command with the czvf options.

Moving Backup to Destination Directory:
Finally, the backup archive is moved from the target_directory_name to the destination_directory_name. The script uses the absolute paths of both directories to ensure the file is moved to the correct location.

Overall, this shell script is a powerful tool for automating the backup process, allowing users to maintain an organized and updated backup of their important files. It's a handy addition to any developer's or system administrator's toolkit.